+++ R/normal_form.R
@@ -12,7 +12,9 @@
 #'     First, it can be a numeric vector of payoffs. Second, it can be a character string of the payoff
 #'    function (e.g., p1 = "x^2 - y"). Third, it can be an R function of payoff.
 #' @param p2 The payoff of Player 2. See the explanation of \code{p1} for detail.
-#' @param byrow A logical value. If \code{TRUE}, payoffs will be lined up by row. Default is FALSE.
+#' @param symmetric A logical value. Set this \code{TRUE} when the payoffs for two players are
+#'     symmetric as in the prisoners' dilemma. Then, p1 is recycled for p2. Default is \code{FALSE}
+#' @param byrow A logical value. If \code{TRUE}, payoffs will be lined up by row. Default is \code{FALSE}.
 #'     Only used when both \code{s1} and \code{s2} are provided.
 #' @param pars A character vector of parameters that are selected by players 1 and 2, respectively.
 #'     Only used when \code{p1} and \code{p2} are specified as a function
@@ -53,6 +55,7 @@
 #'   par2_lim = c(0, 30))
 #'
 #' \dontrun{
+#' ## This causes an error because p1 and p2 are in different forms.
 #' game4 <- normal_form(
 #'   p1 = fx,
 #'   p2 = "-y^2 + (28 - x) * y",
@@ -67,6 +70,7 @@ normal_form <- function(
   s2 = NULL,
   p1,
   p2,
+  symmetric = FALSE,
   byrow = FALSE,
   pars = NULL,
   par1_lim = NULL,
@@ -89,11 +93,14 @@ normal_form <- function(
     n_cols <- length(s2)
     n_cells <- n_rows * n_cols
 
+    if (symmetric) p2 <- p1
+
     if (length(p1) != n_cells) stop("the length of p1 must equal the number of cells.")
     if (length(p2) != n_cells) stop("the length of p2 must equal the number of cells.")
 
     mat1 <- matrix(p1, nrow = n_rows, byrow = byrow)
-    mat2 <- matrix(p2, nrow = n_rows, byrow = byrow)
+    byrow2 <- ifelse(symmetric, !byrow, byrow)
+    mat2 <- matrix(p2, nrow = n_rows, byrow = byrow2)
 
     if (byrow) {
       row <- rep(1:n_rows, each = n_cols)